+++ changelog.md
@@ -4,6 +4,7 @@ Upcoming (TBD)
 Bug Fixes
 ---------
 * Watch command now returns correct time when ran as part of a multi-part query (#1565)
+* Don't diagnose free-entry sections such as `[favorite_queries]` in `--checkup`.
 
 
 1.54.1 (2026/02/17)